Modularize server-side application, and handle master disconnect
On master disconnect, clients redirect and disconnect the socket connection
