+++ src/components/MasterDisconnect.js
@@ -1,0 +1,11 @@
+import React, { Component } from 'react';
+
+const MasterDisconnect = () => {
+  return(<div> 
+              <div className="card well well-lg">
+                <p> The host has disconnected. Please try again. </p>
+              </div>
+          </div>)
+}
+
+export default MasterDisconnect;
